# Improved performance for epee serialization: