--- a/contrib/epee/include/storages/portable_storage_base.h
+++ b/contrib/epee/include/storages/portable_storage_base.h
@@ -111,16 +111,16 @@ namespace epee
         return (t_entry_type*)&(*(m_it++));//fuckoff
       }
 
-      t_entry_type& insert_first_val(const t_entry_type& v)
+      t_entry_type& insert_first_val(t_entry_type&& v)
       {
         m_array.clear();
         m_it = m_array.end();
-        return insert_next_value(v);
+        return insert_next_value(std::move(v));
       }
 
-      t_entry_type& insert_next_value(const t_entry_type& v)
+      t_entry_type& insert_next_value(t_entry_type&& v)
       {
-        m_array.push_back(v);
+        m_array.push_back(std::move(v));
         return m_array.back();
       }