- Removed copy of field names in binary deserialization
  - Removed copy of array values in binary deserialization
  - Removed copy of string values in json deserialization
  - Removed unhelpful allocation in json string value parsing
  - Removed copy of blob data on binary and json serialization

The changes to the binary deserialization and serialization **do** affect the p2p code, which includes the stuff in the key-value serialization code. The changes are fairly straightforward.

I also did some benchmarking:

`GET_BLOCKS_FAST` with 1000 blocks fetched and `pruned == true` (identical to wallet behavior).
Serialization within 1 ms. Deserialization ~4.9% speedup on ryzen 3900x and ~5.6% on i7-7Y75 (mobile). The raw numbers were 3 ms and 6 ms improvement respectively.

`BLOCK_HEADERS_RANGE` with 5000 block headers fetched (not necessarily standard wallet behavior but a larger JSON file). Serialization faster but within 1 ms. Deserialization ~37.7% speed on ryzen 3900x and ~42.6% on i7-7Y75 (mobile). The raw numbers were 17.9 ms and 34.5 ms improvement respectively. Possibly less/more in practice since it depends on the structures and data in use.


Each test was run 100 times before running in a loop 1000 times continuously. Only the last 1000 were used in measurements. The first 100 were an attempt to mitigate weird caching and turbo booting behavior to get consistency.

EDIT: These were run in Linux with Gcc 9.2.